LayoutTests/http/tests/security/xss-DENIED-htmlelelment-with-iframe-proto.html

<html>
<head>
<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cross-frame-access.js"></script>
</head>
<body>
<iframe src="http://localhost:8000/security/resources/iframe-with-element.html" style="">
</iframe>
<pre id="console"></pre>
<script>
description("Tests that making other frame window a prototype doesn't expose that window subframe");
jsTestIsAsync = true;

targetWindow = frames[0];

window.onload = function() {
  shouldThrowErrorName("__proto__ = targetWindow", "TypeError");
  shouldThrowErrorName('targetWindow.myinput', 'SecurityError');
  finishJSTest();
}
</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
